+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change Log for `junit5-system-exit`
 
 ### 2.0.1
-- TBD
+- Bugfix: [[#20]](https://github.com/tginsberg/junit5-system-exit/issues/20): Multiple calls to `System.exit()` do not always report the first exit status code.
 
 ### 2.0.0
 - Remove terminally deprecated `SecurityManager` approach for preventing `System.exit()` calls.